+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # @sikka/hawa
 
+## 0.20.0
+
+### Minor Changes
+
+- - Added `ContactForm`
+  - Fixed `Textarea` placeholder text color
+  - Removed all shadows of `FeedbackForm` and `ContactForm` when `cardless` prop is true
+  - exporting `TextInputType`
+  - General cleanup
+
 ## 0.19.41
 
 ### Patch Changes